+++ src/condor_unit_tests/unit_test_utils.cpp
@@ -175,9 +175,9 @@ ClassAd* get_classad_from_file(){
 	return classad_from_file;
 }
 
-/* Returns true if given floats differ by less than or equal to diff */
-bool floats_close( float one, float two, float diff) {
-	float ftmp = fabs(one) - fabs(two);
+/* Returns true if given double differ by less than or equal to diff */
+bool floats_close( double one, double two, double diff) {
+	double ftmp = fabs(one) - fabs(two);
 	if(fabs(ftmp) <= diff) {
 		return(true);
 	} else {
